requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.2
django-crispy-forms==2.3
djangorestframework==3.16.0
pillow==11.1.0
sqlparse==0.5.3
